requirements.txt
Flask==2.3.3
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3
python-telegram-bot==20.3
requests==2.31.0
schedule==1.2.0
ta==0.10.2
SQLAlchemy==2.0.19
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0